requirements.txt
tensorflow-macos==2.9.2
tensorflow-metal==0.5.0

numpy==1.26.1
opencv-python==4.8.1.78
matplotlib
pyyaml
tqdm